Adding videos of our bot following a line
This lab was tough. Not everything works perfectly or smoothly. You can hear the squeaking in the wheels, this was rough. We learned a lot though. We did this through PID control scheme, but later found out that we could have just used P control and had our hard coded cases (P would have been easier and it was allowed). In the end it was pretty close to doing all cases (even if it was a little whacky).